add some package

+++ 20-install-openshift.sh
@@ -2,10 +2,13 @@
 source 01-config-update.sh
 
 # install the packages for Ansible
-yum -y --enablerepo=epel install pyOpenSSL
 
-curl -o ansible.rpm https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64/ansible-${ANSIBLE_VERSION}-1.el7.ans.noarch.rpm
-yum -y --enablerepo=epel install ansible.rpm
+yum install -y epel-release pyOpenSSL python-cryptography python-lxml
+
+if [ ! -f ansible.rpm ]; then
+    curl -o ansible.rpm https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64/ansible-${ANSIBLE_VERSION}-1.el7.ans.noarch.rpm
+    yum -y --enablerepo=epel install ansible.rpm
+fi
 
 [ ! -d openshift-ansible ] && git clone https://github.com/openshift/openshift-ansible.git -b release-${VERSION} --depth=1
